refactor values and types
